Initial Start respect working hours
