+++ src/PresenceLight/MainWindow.xaml.cs
@@ -126,90 +126,7 @@ namespace PresenceLight
                 property.SetValue(_options, value);
             }
         }
-
-        private async Task LoadSettings()
-        {
-            if (!(await SettingsService.IsFilePresent()))
-            {
-                await SettingsService.SaveSettings(_options);
-            }
-
-            Config = await SettingsService.LoadSettings();
-
-            if (string.IsNullOrEmpty(Config.RedirectUri))
-            {
-                await SettingsService.DeleteSettings();
-                await SettingsService.SaveSettings(_options);
-            }
-            if (Config.LightSettings.UseWorkingHours)
-            {
-                pnlWorkingHours.Visibility = Visibility.Visible;
-                SyncOptions();
-            }
-            else
-            {
-                pnlWorkingHours.Visibility = Visibility.Collapsed;
-                SyncOptions();
-            }
-
-            if (Config.LightSettings.Hue.IsPhillipsHueEnabled)
-            {
-                pnlPhillips.Visibility = Visibility.Visible;
-                SyncOptions();
-            }
-            else
-            {
-                pnlPhillips.Visibility = Visibility.Collapsed;
-            }
-
-            if (Config.LightSettings.Yeelight.IsYeelightEnabled)
-            {
-                pnlYeelight.Visibility = Visibility.Visible;
-                SyncOptions();
-            }
-            else
-            {
-                pnlYeelight.Visibility = Visibility.Collapsed;
-            }
-
-            if (Config.LightSettings.LIFX.IsLIFXEnabled)
-            {
-                getTokenLink.Visibility = Visibility.Visible;
-                pnlLIFX.Visibility = Visibility.Visible;
-
-                SyncOptions();
-            }
-            else
-            {
-                getTokenLink.Visibility = Visibility.Collapsed;
-                pnlLIFX.Visibility = Visibility.Collapsed;
-            }
-
-            if (Config.LightSettings.Custom.IsCustomApiEnabled)
-            {
-                pnlCustomApi.Visibility = Visibility.Visible;
-                customApiAvailableMethod.SelectedValue = Config.LightSettings.Custom.CustomApiAvailableMethod;
-                customApiBusyMethod.SelectedValue = Config.LightSettings.Custom.CustomApiBusyMethod;
-                customApiBeRightBackMethod.SelectedValue = Config.LightSettings.Custom.CustomApiBeRightBackMethod;
-                customApiAwayMethod.SelectedValue = Config.LightSettings.Custom.CustomApiAwayMethod;
-                customApiDoNotDisturbMethod.SelectedValue = Config.LightSettings.Custom.CustomApiDoNotDisturbMethod;
-                customApiOfflineMethod.SelectedValue = Config.LightSettings.Custom.CustomApiOfflineMethod;
-                customApiOffMethod.SelectedValue = Config.LightSettings.Custom.CustomApiOffMethod;
-                SyncOptions();
-            }
-            else
-            {
-                pnlCustomApi.Visibility = Visibility.Collapsed;
-            }
-            if (!string.IsNullOrEmpty(Config.LightSettings.LIFX.LIFXClientId) && !(string.IsNullOrEmpty(Config.LightSettings.LIFX.LIFXClientSecret)))
-            {
-                getTokenLink.Visibility = Visibility.Visible;
-            }
-            else
-            {
-                getTokenLink.Visibility = Visibility.Collapsed;
-            }
-        }
+     
         #endregion
 
         #region Profile Panel
@@ -240,31 +157,38 @@ namespace PresenceLight
             signInPanel.Visibility = Visibility.Collapsed;
             lblTheme.Visibility = Visibility.Collapsed;
             loadingPanel.Visibility = Visibility.Visible;
-            var (profile, presence) = await System.Threading.Tasks.Task.Run(() => GetBatchContent());
-            var photo = await System.Threading.Tasks.Task.Run(() => GetPhoto());
 
-            if (photo == null)
+            if (Config.LightSettings.SyncLights)
             {
-                MapUI(presence, profile, new BitmapImage(new Uri("pack://application:,,,/PresenceLight;component/images/UnknownProfile.png")));
-            }
-            else
-            {
-                MapUI(presence, profile, LoadImage(photo));
-            }
+                if (!Config.LightSettings.UseWorkingHours || (Config.LightSettings.UseWorkingHours && IsInWorkingHours()))
+                {
+                    var (profile, presence) = await System.Threading.Tasks.Task.Run(() => GetBatchContent());
+                    var photo = await System.Threading.Tasks.Task.Run(() => GetPhoto());
 
-            if (lightMode == "Graph")
-            {
-                await SetColor(presence.Availability);
+                    if (photo == null)
+                    {
+                        MapUI(presence, profile, new BitmapImage(new Uri("pack://application:,,,/PresenceLight;component/images/UnknownProfile.png")));
+                    }
+                    else
+                    {
+                        MapUI(presence, profile, LoadImage(photo));
+                    }
+
+                    if (lightMode == "Graph")
+                    {
+                        await SetColor(presence.Availability);
+                    }
+                    loadingPanel.Visibility = Visibility.Collapsed;
+                    this.signInPanel.Visibility = Visibility.Collapsed;
+                    hueIpAddress.IsEnabled = false;
+
+                    dataPanel.Visibility = Visibility.Visible;
+                    await SettingsService.SaveSettings(Config);
+
+                    turnOffButton.Visibility = Visibility.Visible;
+                    turnOnButton.Visibility = Visibility.Collapsed;
+                }
             }
-            loadingPanel.Visibility = Visibility.Collapsed;
-            this.signInPanel.Visibility = Visibility.Collapsed;
-            hueIpAddress.IsEnabled = false;
-
-            dataPanel.Visibility = Visibility.Visible;
-            await SettingsService.SaveSettings(Config);
-
-            turnOffButton.Visibility = Visibility.Visible;
-            turnOnButton.Visibility = Visibility.Collapsed;
 
             await InteractWithLights();
         }
@@ -482,142 +406,7 @@ namespace PresenceLight
 
             return (User: user, Presence: presence);
         }
-        #endregion
-
-        #region Settings Panel
-        private async void SaveSettings_Click(object sender, RoutedEventArgs e)
-        {
-            btnSettings.IsEnabled = false;
-            if (Transparent.IsChecked == true)
-            {
-                Config.IconType = "Transparent";
-            }
-            else
-            {
-                Config.IconType = "White";
-            }
-
-            CheckAAD();
-            Config.LightSettings.DefaultBrightness = Convert.ToInt32(brightness.Value);
-
-            SetWorkingDays();
-
-            SyncOptions();
-            await SettingsService.SaveSettings(Config);
-            lblSettingSaved.Visibility = Visibility.Visible;
-            btnSettings.IsEnabled = true;
-        }
-
-        private void SetWorkingDays()
-        {
-            List<string> days = new List<string>();
-
-            if (Monday.IsChecked.Value)
-            {
-                days.Add("Monday");
-            }
-
-            if (Tuesday.IsChecked.Value)
-            {
-                days.Add("Tuesday");
-            }
-
-            if (Wednesday.IsChecked.Value)
-            {
-                days.Add("Wednesday");
-            }
-
-            if (Thursday.IsChecked.Value)
-            {
-                days.Add("Thursday");
-            }
-
-            if (Friday.IsChecked.Value)
-            {
-                days.Add("Friday");
-            }
-
-            if (Saturday.IsChecked.Value)
-            {
-                days.Add("Saturday");
-            }
-
-            if (Sunday.IsChecked.Value)
-            {
-                days.Add("Sunday");
-            }
-
-            Config.LightSettings.WorkingDays = string.Join("|", days);
-        }
-
-        private void CheckAAD()
-        {
-            Regex r = new Regex(@"^(\{){0,1}[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}(\}){0,1}$");
-            if (string.IsNullOrEmpty(Config.ClientId) || string.IsNullOrEmpty(Config.RedirectUri) || !r.IsMatch(Config.ClientId))
-            {
-                configErrorPanel.Visibility = Visibility.Visible;
-                dataPanel.Visibility = Visibility.Hidden;
-                signInPanel.Visibility = Visibility.Hidden;
-                return;
-            }
-
-            SyncOptions();
-
-            configErrorPanel.Visibility = Visibility.Hidden;
-
-            if (dataPanel.Visibility != Visibility.Visible)
-            {
-                signInPanel.Visibility = Visibility.Visible;
-            }
-
-            if (_graphServiceClient == null)
-            {
-                _graphServiceClient = _graphservice.GetAuthenticatedGraphClient(typeof(WPFAuthorizationProvider));
-            }
-        }
-        #endregion
-
-        private void PopulateWorkingDays()
-        {
-            if (!string.IsNullOrEmpty(Config.LightSettings.WorkingDays))
-            {
-
-                if (Config.LightSettings.WorkingDays.Contains("Monday"))
-                {
-                    Monday.IsChecked = true;
-                }
-
-                if (Config.LightSettings.WorkingDays.Contains("Tuesday"))
-                {
-                    Tuesday.IsChecked = true;
-                }
-
-                if (Config.LightSettings.WorkingDays.Contains("Wednesday"))
-                {
-                    Wednesday.IsChecked = true;
-                }
-
-                if (Config.LightSettings.WorkingDays.Contains("Thursday"))
-                {
-                    Thursday.IsChecked = true;
-                }
-
-                if (Config.LightSettings.WorkingDays.Contains("Friday"))
-                {
-                    Friday.IsChecked = true;
-                }
-
-                if (Config.LightSettings.WorkingDays.Contains("Saturday"))
-                {
-                    Saturday.IsChecked = true;
-                }
-
-                if (Config.LightSettings.WorkingDays.Contains("Sunday"))
-                {
-                    Sunday.IsChecked = true;
-                }
-            }
-        }
+        #endregion      
 
 
         private void TabControl_SelectionChanged(object sender, System.Windows.Controls.SelectionChangedEventArgs e)
@@ -627,60 +416,7 @@ namespace PresenceLight
             lblSettingSaved.Visibility = Visibility.Collapsed;
         }
 
-        private async void cbSyncLights(object sender, RoutedEventArgs e)
-        {
-            if (!Config.LightSettings.SyncLights)
-            {
-                await SetColor("Off");
-                turnOffButton.Visibility = Visibility.Collapsed;
-                turnOnButton.Visibility = Visibility.Visible;
-            }
-
-            SyncOptions();
-            await SettingsService.SaveSettings(Config);
-            e.Handled = true;
-        }
-
-        private async void cbUseDefaultBrightnessChanged(object sender, RoutedEventArgs e)
-        {
-            if (Config.LightSettings.UseDefaultBrightness)
-            {
-                pnlDefaultBrightness.Visibility = Visibility.Visible;
-            }
-            else
-            {
-                pnlDefaultBrightness.Visibility = Visibility.Collapsed;
-            }
-
-            SyncOptions();
-            await SettingsService.SaveSettings(Config);
-            e.Handled = true;
-        }
-
-        private void cbUseWorkingHoursChanged(object sender, RoutedEventArgs e)
-        {
-            if (!string.IsNullOrEmpty(Config.LightSettings.WorkingHoursStartTime))
-            {
-                Config.LightSettings.WorkingHoursStartTime = DateTime.Parse(Config.LightSettings.WorkingHoursStartTime).TimeOfDay.ToString();
-            }
-
-            if (!string.IsNullOrEmpty(Config.LightSettings.WorkingHoursEndTime))
-            {
-                Config.LightSettings.WorkingHoursEndTime = DateTime.Parse(Config.LightSettings.WorkingHoursEndTime).TimeOfDay.ToString();
-            }
-
-            if (Config.LightSettings.UseWorkingHours)
-            {
-                pnlWorkingHours.Visibility = Visibility.Visible;
-            }
-            else
-            {
-                pnlWorkingHours.Visibility = Visibility.Collapsed;
-            }
-
-            SyncOptions();
-            e.Handled = true;
-        }
+       
 
         #region Tray Methods
 
@@ -699,6 +435,7 @@ namespace PresenceLight
                 this.WindowState = this.lastWindowState;
             }
         }
+
         private void OnOpenClick(object sender, RoutedEventArgs e)
         {
             this.Show();
@@ -739,7 +476,6 @@ namespace PresenceLight
             this.WindowState = this.lastWindowState;
         }
 
-
         private async void OnExitClick(object sender, RoutedEventArgs e)
         {
             if (!string.IsNullOrEmpty(Config.LightSettings.Hue.HueApiKey) && !string.IsNullOrEmpty(Config.LightSettings.Hue.HueIpAddress) && !string.IsNullOrEmpty(Config.LightSettings.Hue.SelectedHueLightId))
@@ -755,6 +491,7 @@ namespace PresenceLight
             await SettingsService.SaveSettings(Config);
             System.Windows.Application.Current.Shutdown();
         }
+
         private async void Current_SessionEnding(object sender, SessionEndingCancelEventArgs e)
         {
             if (!string.IsNullOrEmpty(Config.LightSettings.Hue.HueApiKey) && !string.IsNullOrEmpty(Config.LightSettings.Hue.HueIpAddress) && !string.IsNullOrEmpty(Config.LightSettings.Hue.SelectedHueLightId))
@@ -775,52 +512,9 @@ namespace PresenceLight
 
             await SettingsService.SaveSettings(Config);
         }
+
         #endregion
-
-        bool IsInWorkingHours()
-        {
-            if (string.IsNullOrEmpty(Config.LightSettings.WorkingHoursStartTime) || string.IsNullOrEmpty(Config.LightSettings.WorkingHoursEndTime) || string.IsNullOrEmpty(Config.LightSettings.WorkingDays))
-            {
-                return false;
-            }
-
-            if (!Config.LightSettings.WorkingDays.Contains(DateTime.Now.DayOfWeek.ToString()))
-            {
-                return false;
-            }
-
-            // convert datetime to a TimeSpan
-            bool validStart = TimeSpan.TryParse(Config.LightSettings.WorkingHoursStartTime, out TimeSpan start);
-            bool validEnd = TimeSpan.TryParse(Config.LightSettings.WorkingHoursEndTime, out TimeSpan end);
-            if (!validEnd || !validStart)
-            {
-                return false;
-            }
-
-            TimeSpan now = DateTime.Now.TimeOfDay;
-            // see if start comes before end
-            if (start < end)
-                return start <= now && now <= end;
-            // start is after end, so do the inverse comparison
-            return !(end < now && now < start);
-        }
-
-        private void time_ValueChanged(object sender, RoutedPropertyChangedEventArgs<object> e)
-        {
-            if (!string.IsNullOrEmpty(Config.LightSettings.WorkingHoursStartTime))
-            {
-                Config.LightSettings.WorkingHoursStartTime = DateTime.Parse(Config.LightSettings.WorkingHoursStartTime).TimeOfDay.ToString();
-            }
-
-            if (!string.IsNullOrEmpty(Config.LightSettings.WorkingHoursEndTime))
-            {
-                Config.LightSettings.WorkingHoursEndTime = DateTime.Parse(Config.LightSettings.WorkingHoursEndTime).TimeOfDay.ToString();
-            }
-
-            SyncOptions();
-            e.Handled = true;
-        }
-
+        
         private async Task InteractWithLights()
         {
             while (true)
